add makepage and move main.dart
